refactor(main): move http-server and transpor out of main
